tests/requirements-test.txt
# Test dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-timeout==2.2.0
pytest-env==1.1.1
pytest-sugar==1.0.0
pytest-html==4.1.1
pytest-metadata==3.1.0
pytest-ordering==0.6
pytest-randomly==3.15.0
pytest-repeat==0.9.1
pytest-rerunfailures==12.0
pytest-selenium==4.0.2
pytest-socket==0.6.0
pytest-watch==4.2.0

# Coverage
coverage==7.3.2
coverage-badge==1.1.0

# Mocking
mock==5.1.0
freezegun==1.4.0
responses==0.24.1
vcrpy==4.2.1

# Type checking
mypy==1.7.1
types-mock==5.1.0.20240106
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240106
types-setuptools==69.0.0.20240106
types-urllib3==1.26.25.14

# Code quality
pylint==3.0.3
pylint-pytest==1.1.2
pylint-pytest-plugin==0.1.0
pylint-django==2.5.5
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-celery==0.3
pylint-airflow==0.1.0
pylint-pandas==0.2.3
pylint-numpy==0.1.0
pylint-scipy==0.1.0
pylint-tensorflow==0.1.0
pylint-torch==0.1.0
pylint-keras==0.1.0
pylint-sklearn==0.1.0

# Security
bandit==1.7.5
safety==2.3.5
dlint==0.14.0
yara-python==4.3.1
python-nmap==0.7.1
paramiko==3.4.0
cryptography==41.0.7

# Performance
locust==2.20.1
locust-plugins==3.4.1
locust-influxdb-listener==0.1.0
locust-prometheus-exporter==0.1.0
locust-grafana-dashboard==0.1.0
locust-elasticsearch-listener==0.1.0
locust-kafka-listener==0.1.0
locust-redis-listener==0.1.0
locust-mongodb-listener==0.1.0
locust-postgresql-listener==0.1.0
locust-mysql-listener==0.1.0
locust-sqlite-listener==0.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2
sphinx-copybutton==0.5.2
sphinx-tabs==3.4.4
sphinx-panels==0.6.0
sphinx-prompt==1.5.0
sphinx-issues==3.0.1
sphinx-github-role==1.0.0
sphinx-js==3.0.1
sphinx-jsonschema==1.21.0
sphinx-markdown-tables==0.0.17
sphinx-multiversion==0.2.4
sphinx-notfound-page==0.8.3
sphinx-press-theme==0.8.0
sphinx-sitemap==2.2.0
sphinx-substitution-extensions==2023.1
sphinx-toolbox==3.5.0
sphinxcontrib-bibtex==2.6.0
sphinxcontrib-mermaid==0.9.1
sphinxcontrib-plantuml==0.25
sphinxcontrib-programoutput==0.17
sphinxcontrib-spelling==7.10.0
sphinxcontrib-svg2pdfconverter==1.2.2
sphinxcontrib-websupport==1.2.4
sphinxcontrib-youtube==1.0.1

# Development
ipython==8.20.0
jupyter==1.0.0
jupyterlab==4.0.9
notebook==7.0.6
black==23.12.1
isort==5.13.2
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-bugbear==23.12.2
flake8-comprehensions==3.14.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-variables-names==0.0.5
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-variables-names==0.0.5
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-variables-names==0.0.5

# Debugging
ipdb==0.13.13
debugpy==1.8.0
memory-profiler==0.61.0
line-profiler==4.1.1
objgraph==3.6.0
py-spy==0.3.14
pyinstrument==4.6.1
pyflame==1.7.0
py-spy==0.3.14
py-spy-viewer==0.1.0
py-spy-record==0.1.0
py-spy-top==0.1.0
py-spy-dump==0.1.0
py-spy-record==0.1.0
py-spy-top==0.1.0
py-spy-dump==0.1.0

# CI/CD
tox==4.12.1
tox-gh-actions==2.12.0
tox-gh==1.4.0
tox-travis==0.12
tox-docker==4.0.0
tox-conda==0.4.0
tox-envfile==3.1.0
tox-factor==0.1.0
tox-pip-extensions==1.6.0
tox-pip-sync==0.1.0
tox-pipenv==0.1.0
tox-poetry==0.1.0
tox-pyenv==0.1.0
tox-virtualenv==0.1.0
tox-wheel==0.1.0
tox-workflow==0.1.0
tox-workflow-gh==0.1.0
tox-workflow-travis==0.1.0
tox-workflow-docker==0.1.0
tox-workflow-conda==0.1.0
tox-workflow-envfile==0.1.0
tox-workflow-factor==0.1.0
tox-workflow-pip-extensions==0.1.0
tox-workflow-pip-sync==0.1.0
tox-workflow-pipenv==0.1.0
tox-workflow-poetry==0.1.0
tox-workflow-pyenv==0.1.0
tox-workflow-virtualenv==0.1.0
tox-workflow-wheel==0.1.0 